Let autogenerate compare only specific tables
Adds an additional argument to compare_metadata(), that allows to
specify names of the tables, which should be compared (differences
in other tables are ignored). A caller might pass either a sequence
of table names or a callable which accepts a table name and returns
True for if the table should be compared.

This is modeled after reflect() method of SQLAlchemy MetaData class.