Add include_branch arg to 'current' function
This commit updates the 'current' function in alembic/command.py by
adding include_branches argument.  With this in place, we then pass
include_branch to the 'cmd_format' call inside 'current'.

This is being implemented to help solve [1].  OpenStack Neutron uses
multiple alembic branches and as a deployer I would like to use
'neutron-db-manage current' (which is leveraging alembic's 'current')
to determine which revision is associated with which branch so I can
selectively apply migrations for each branch.  This is necessary as one
branch can have migrations run while neutron-server is running while the
other should be run with neutron-server stopped.

I am not familiar w/ alembic, so please let me know if this change is
inappropriate or if there is a better way to achieve the desired
result.

[1] https://bugs.launchpad.net/neutron/+bug/1488021